# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_model_address.Address:
    properties:
      apartment:
        type: string
      city:
        type: string
      comment:
        type: string
      house:
        type: string
      street:
        type: string
    type: object
  github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_model_money.Money:
    type: object
  github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorDetails:
    properties:
      err:
        type: string
      message:
        type: string
    type: object
  ? github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse
  : properties:
      error:
        $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorDetails'
    type: object
  internal_handler_catalog.AddressDTO:
    properties:
      apartment:
        type: string
      city:
        type: string
      house:
        type: string
      street:
        type: string
    type: object
  internal_handler_catalog.CategoryDTO:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  internal_handler_catalog.MenuItemDTO:
    properties:
      available:
        type: boolean
      category:
        $ref: '#/definitions/internal_handler_catalog.CategoryDTO'
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: string
      restaurant_id:
        type: string
    type: object
  internal_handler_catalog.MenuItemsDTO:
    properties:
      menu_items:
        items:
          $ref: '#/definitions/internal_handler_catalog.MenuItemDTO'
        type: array
    type: object
  internal_handler_catalog.RestaurantDTO:
    properties:
      address:
        $ref: '#/definitions/internal_handler_catalog.AddressDTO'
      description:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
  internal_handler_catalog.RestaurantsDTO:
    properties:
      restaurants:
        items:
          $ref: '#/definitions/internal_handler_catalog.RestaurantDTO'
        type: array
    type: object
  internal_handler_orders.AddressDTO:
    properties:
      apartment:
        type: string
      city:
        type: string
      comment:
        type: string
      house:
        type: string
      street:
        type: string
    type: object
  internal_handler_orders.CreateOrderDTORequest:
    properties:
      comment:
        type: string
      delivery_address:
        $ref: '#/definitions/internal_handler_orders.AddressDTO'
      order_items:
        items:
          $ref: '#/definitions/internal_handler_orders.CreateOrderItemDTORequest'
        type: array
      restaurant_id:
        type: string
    type: object
  internal_handler_orders.CreateOrderDTOResponse:
    properties:
      estimated_delivery_time:
        type: string
      id:
        type: string
      status:
        type: string
      total_amount:
        type: string
    type: object
  internal_handler_orders.CreateOrderItemDTORequest:
    properties:
      menu_item_id:
        type: string
      quantity:
        type: integer
    type: object
  internal_handler_orders.OrderDTO:
    properties:
      comment:
        type: string
      created_at:
        type: string
      delivery_address:
        $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_model_address.Address'
      id:
        type: string
      order_items:
        items:
          $ref: '#/definitions/internal_handler_orders.OrderItemDTO'
        type: array
      restaurant_id:
        type: string
      status:
        type: string
      total_amount:
        $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_model_money.Money'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  internal_handler_orders.OrderItemDTO:
    properties:
      id:
        type: string
      menu_item_id:
        type: string
      name:
        type: string
      order_id:
        type: string
      price:
        $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_model_money.Money'
      quantity:
        type: integer
    type: object
  internal_handler_orders.OrdersDTO:
    properties:
      orders:
        items:
          $ref: '#/definitions/internal_handler_orders.OrderDTO'
        type: array
    type: object
  internal_handler_orders.RejectOrderDTO:
    properties:
      reason:
        type: string
    type: object
  internal_handler_orders.UpdateMenuDTO:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_handler_orders.UpdateMenuItemDTO'
        type: array
    type: object
  internal_handler_orders.UpdateMenuItemDTO:
    properties:
      available:
        type: boolean
      category_id:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: string
    type: object
  internal_handler_orders.UpdateOrderStatusDTO:
    properties:
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: MVP API агрегатора доставки еды «Авито.Кухня».
  title: Avito Kitchen API
  version: 1.0.0
paths:
  /categories:
    get:
      consumes:
      - application/json
      description: Возвращает список зарегистрированных в сервисе категорий
      produces:
      - application/json
      responses:
        "200":
          description: Информация о заведении
          schema:
            $ref: '#/definitions/internal_handler_catalog.RestaurantDTO'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
      summary: Список категорий
      tags:
      - catalog
  /orders:
    post:
      consumes:
      - application/json
      description: Создает заказ
      parameters:
      - description: UUID пользователя. Если не передан - заказ анонимный. Замена
          авторизации в MVP.
        format: uuid
        in: header
        name: X-User-ID
        type: string
      - description: Тело запроса
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/internal_handler_orders.CreateOrderDTORequest'
      produces:
      - application/json
      responses:
        "201":
          description: Созданный заказ
          schema:
            $ref: '#/definitions/internal_handler_orders.CreateOrderDTOResponse'
        "400":
          description: Невалидный запрос
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "404":
          description: Не найдено
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "409":
          description: Конфликт
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
      summary: Создать заказ
      tags:
      - orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает подробную информацию о заказе
      parameters:
      - description: Идентификатор заказа
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация о заказе
          schema:
            $ref: '#/definitions/internal_handler_orders.OrderDTO'
        "400":
          description: Невалидный запрос
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "404":
          description: Заказ не найден
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
      summary: Получить информацию о заказе
      tags:
      - orders
  /restaurants:
    get:
      consumes:
      - application/json
      description: Возвращает список активных заведений
      produces:
      - application/json
      responses:
        "200":
          description: Список зведений
          schema:
            $ref: '#/definitions/internal_handler_catalog.RestaurantsDTO'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
      summary: Список активных заведений
      tags:
      - catalog
  /restaurants/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает информацию о заведении
      parameters:
      - description: Идентификатор заведения
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация о заведении
          schema:
            $ref: '#/definitions/internal_handler_catalog.RestaurantDTO'
        "400":
          description: Невалидный запрос (неправильный формат UUID ресторана)
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "404":
          description: Заведение не найдено
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
      summary: Детали заведения
      tags:
      - catalog
  /restaurants/{id}/menu:
    get:
      consumes:
      - application/json
      description: Возвращает меню заведения
      parameters:
      - description: Идентификатор заведения
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Меню заведения
          schema:
            $ref: '#/definitions/internal_handler_catalog.MenuItemsDTO'
        "400":
          description: Невалидный запрос (неправильный формат UUID ресторана)
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "404":
          description: Заведение не найдено
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
      summary: Меню заведения
      tags:
      - catalog
  /restaurants/menu:
    put:
      consumes:
      - application/json
      description: Обновляет меню заведения целиком и возвращает обновленное меню
      parameters:
      - description: UUID ресторана. Замена авторизации в MVP.
        format: uuid
        in: header
        name: X-Restaurant-ID
        required: true
        type: string
      - description: Тело запроса
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/internal_handler_orders.UpdateMenuDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленное меню
          schema:
            $ref: '#/definitions/internal_handler_orders.UpdateMenuDTO'
        "400":
          description: Невалидный запрос
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "404":
          description: Заведение не найдено
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
      summary: Обновить меню целиком
      tags:
      - restaurant
  /restaurants/orders:
    get:
      consumes:
      - application/json
      description: Возвращает активные заказы заведения
      parameters:
      - description: UUID ресторана. UUID заведения, от имени которого выполняется
          запрос (замена авторизации в MVP).
        format: uuid
        in: header
        name: X-Restaurant-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация об активных заказах
          schema:
            $ref: '#/definitions/internal_handler_orders.OrdersDTO'
        "400":
          description: Невалидный запрос (неверный формат UUID)
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "404":
          description: Ресторан не найден
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
      summary: Заказы заведения (pending + активные)
      tags:
      - restaurant
  /restaurants/orders/{id}/accept:
    post:
      consumes:
      - application/json
      description: Принимает заказ
      parameters:
      - description: UUID ресторана. Замена авторизации в MVP.
        format: uuid
        in: header
        name: X-Restaurant-ID
        required: true
        type: string
      - description: Идентификатор заказа
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Принятый заказ
          schema:
            $ref: '#/definitions/internal_handler_orders.OrderDTO'
        "400":
          description: Невалидный запрос
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "403":
          description: Заказ не принадлежит данному ресторану
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "404":
          description: Заказ не найден
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "409":
          description: Конфликт при обновлении статуса заказа
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
      summary: Принять заказ
      tags:
      - restaurant
  /restaurants/orders/{id}/reject:
    post:
      consumes:
      - application/json
      description: Отклоняет заказ с объяснением причины
      parameters:
      - description: UUID ресторана. Замена авторизации в MVP.
        format: uuid
        in: header
        name: X-Restaurant-ID
        required: true
        type: string
      - description: Идентификатор заказа
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Тело запроса
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/internal_handler_orders.RejectOrderDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Отклоненный заказ
          schema:
            $ref: '#/definitions/internal_handler_orders.OrderDTO'
        "400":
          description: Невалидный запрос
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "403":
          description: Заказ не принадлежит данному ресторану
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "404":
          description: Заказ не найден
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "409":
          description: Конфликт при обновлении статуса заказа
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
      summary: Отклонить заказ
      tags:
      - restaurant
  /restaurants/orders/{id}/status:
    patch:
      consumes:
      - application/json
      description: Обновляет статус заказа на указанный
      parameters:
      - description: UUID ресторана. Замена авторизации в MVP.
        format: uuid
        in: header
        name: X-Restaurant-ID
        required: true
        type: string
      - description: Идентификатор заказа
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Тело запроса
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/internal_handler_orders.UpdateOrderStatusDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Отклоненный заказ
          schema:
            $ref: '#/definitions/internal_handler_orders.OrderDTO'
        "400":
          description: Невалидный запрос
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "403":
          description: Заказ не принадлежит данному ресторану
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "404":
          description: Заказ не найден
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "409":
          description: Конфликт при обновлении статуса заказа (неверно задан следующий
            статус обновления)
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_talense-tasks_backend-trainee-assignment-autumn-2026-markgredasov-5b2b61ca_internal_server_http_response.ErrorResponse'
      summary: Обновить статус заказа
      tags:
      - restaurant
swagger: "2.0"
tags:
- description: Каталог заведений, меню, категорий (пользовательская часть).
  name: catalog
- description: Заказы (пользовательская часть).
  name: orders
- description: API для заведений.
  name: restaurant